+++ engine/include/fury/fury.hpp
@@ -36,7 +36,7 @@ namespace fury {
 
 inline constexpr const char* engine_name() { return "Fury"; }
 inline constexpr int version_major() { return 2; }
-inline constexpr int version_minor() { return 3; }
+inline constexpr int version_minor() { return 4; }
 inline constexpr int version_patch() { return 0; }
 
 }  // namespace fury